requirements.txt
aiohttp==3.8.1
celery==4.4.7
fastapi[all]==0.89.1
fastapi-pagination==0.11.4
flower==0.9.7
gunicorn==20.1.0
Authlib==1.2.0
SQLAlchemy==2.0.0
pillow==10.3.0
psycopg2==2.9.5
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pytest==7.2.1
pytest-asyncio==0.19.0
alembic==1.9.2
python-json-logger==2.0.7
python-multipart==0.0.5
redis==3.5.3
requests==2.28.2
slowapi==0.1.7
ruff==0.0.278
black==22.10.0
starlette==0.22.0
prometheus-fastapi-instrumentator